services/ocr-service/app/core/image_processor.py
"""
Image Processor

Image preprocessing and enhancement functionality using OpenCV.
"""

from typing import Tuple, Optional, List, Dict, Any
import numpy as np
from PIL import Image
import cv2


class ImageProcessor:
    """
    Image processor for preprocessing invoice images before OCR.

    Handles image enhancement, noise reduction, and geometric corrections.
    """

    def __init__(self):
        """Initialize image processor."""
        pass

    def preprocess_image(
        self,
        image: np.ndarray,
        operations: Optional[List[str]] = None
    ) -> np.ndarray:
        """
        Apply preprocessing operations to image.

        Args:
            image: Input image as numpy array
            operations: List of preprocessing operations to apply

        Returns:
            Preprocessed image
        """
        # TODO: Implement image preprocessing pipeline
        pass

    def enhance_contrast(
        self,
        image: np.ndarray,
        alpha: float = 1.0,
        beta: int = 0
    ) -> np.ndarray:
        """
        Enhance image contrast.

        Args:
            image: Input image
            alpha: Contrast control (1.0-3.0)
            beta: Brightness control (0-100)

        Returns:
            Contrast-enhanced image
        """
        # TODO: Implement contrast enhancement
        pass

    def denoise_image(
        self,
        image: np.ndarray,
        filter_strength: int = 10
    ) -> np.ndarray:
        """
        Remove noise from image.

        Args:
            image: Input image
            filter_strength: Denoising filter strength

        Returns:
            Denoised image
        """
        # TODO: Implement noise reduction
        pass

    def deskew_image(
        self,
        image: np.ndarray,
        angle_threshold: float = 15.0
    ) -> np.ndarray:
        """
        Correct image skew/rotation.

        Args:
            image: Input image
            angle_threshold: Maximum angle to correct (degrees)

        Returns:
            Deskewed image
        """
        # TODO: Implement deskewing
        pass

    def detect_document_edges(
        self,
        image: np.ndarray
    ) -> Optional[np.ndarray]:
        """
        Detect document edges for perspective correction.

        Args:
            image: Input image

        Returns:
            Array of corner points or None if not detected
        """
        # TODO: Implement edge detection
        pass

    def correct_perspective(
        self,
        image: np.ndarray,
        corners: Optional[np.ndarray] = None
    ) -> np.ndarray:
        """
        Apply perspective correction to image.

        Args:
            image: Input image
            corners: Document corner points (auto-detect if None)

        Returns:
            Perspective-corrected image
        """
        # TODO: Implement perspective correction
        pass

    def resize_image(
        self,
        image: np.ndarray,
        max_width: int = 2048,
        max_height: int = 2048,
        maintain_aspect: bool = True
    ) -> np.ndarray:
        """
        Resize image to optimal dimensions.

        Args:
            image: Input image
            max_width: Maximum width
            max_height: Maximum height
            maintain_aspect: Whether to maintain aspect ratio

        Returns:
            Resized image
        """
        # TODO: Implement image resizing
        pass

    def convert_to_grayscale(
        self,
        image: np.ndarray
    ) -> np.ndarray:
        """
        Convert image to grayscale.

        Args:
            image: Input color image

        Returns:
            Grayscale image
        """
        # TODO: Implement grayscale conversion
        pass

    def apply_threshold(
        self,
        image: np.ndarray,
        threshold_type: str = "adaptive"
    ) -> np.ndarray:
        """
        Apply thresholding to create binary image.

        Args:
            image: Input grayscale image
            threshold_type: Type of thresholding ("adaptive", "otsu", "manual")

        Returns:
            Binary image
        """
        # TODO: Implement thresholding
        pass

    def get_image_quality_score(
        self,
        image: np.ndarray
    ) -> Dict[str, float]:
        """
        Assess image quality metrics.

        Args:
            image: Input image

        Returns:
            Dictionary with quality metrics (sharpness, brightness, contrast)
        """
        # TODO: Implement quality assessment
        pass