services/ocr-service/app/core/ocr_engine.py
"""
OCR Engine

Main OCR processing engine using PaddleOCR for Thai text extraction.
"""

from typing import List, Dict, Any, Optional, Tuple
import numpy as np
from PIL import Image


class OCREngine:
    """
    OCR Engine for processing Thai invoices using PaddleOCR.

    Handles text extraction, confidence scoring, and result formatting.
    """

    def __init__(self, language: str = "th", use_gpu: bool = False):
        """
        Initialize OCR engine.

        Args:
            language: Language code for OCR (default: "th" for Thai)
            use_gpu: Whether to use GPU acceleration
        """
        self.language = language
        self.use_gpu = use_gpu
        self._ocr_instance = None

    def initialize(self) -> None:
        """
        Initialize PaddleOCR instance.

        This should be called once before processing images.
        """
        # TODO: Initialize PaddleOCR with proper configuration
        pass

    def extract_text(
        self,
        image: np.ndarray,
        confidence_threshold: float = 0.5
    ) -> List[Dict[str, Any]]:
        """
        Extract text from image with confidence scores.

        Args:
            image: Input image as numpy array
            confidence_threshold: Minimum confidence score for text detection

        Returns:
            List of text extraction results with bounding boxes and confidence
        """
        # TODO: Implement text extraction logic
        pass

    def extract_invoice_fields(
        self,
        image: np.ndarray
    ) -> Dict[str, Any]:
        """
        Extract structured invoice fields from image.

        Args:
            image: Input invoice image

        Returns:
            Dictionary containing extracted invoice fields
        """
        # TODO: Implement invoice field extraction
        pass

    def get_text_regions(
        self,
        image: np.ndarray
    ) -> List[Tuple[List[List[int]], str, float]]:
        """
        Get text regions with bounding boxes.

        Args:
            image: Input image

        Returns:
            List of (bounding_box, text, confidence) tuples
        """
        # TODO: Implement text region detection
        pass

    def calculate_overall_confidence(
        self,
        results: List[Dict[str, Any]]
    ) -> float:
        """
        Calculate overall confidence score for OCR results.

        Args:
            results: List of OCR extraction results

        Returns:
            Overall confidence score (0.0 to 1.0)
        """
        # TODO: Implement confidence calculation
        pass

    def is_initialized(self) -> bool:
        """
        Check if OCR engine is initialized.

        Returns:
            True if engine is ready for processing
        """
        # TODO: Check initialization status
        pass

    def cleanup(self) -> None:
        """
        Clean up OCR engine resources.
        """
        # TODO: Implement cleanup logic
        pass